Split test cases into 2 separate files
Moved all the ’events’ test cases into a separate file.

Updated helperFns.js to have DRY code.
